LayoutTests/svg/animations/animate-color-fill-from-by-expected.txt

SVG 1.1 dynamic animation tests

Tests animation on 'currentColor'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Initial condition:
PASS colorComponents[1] is 208
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0

Half-time condition:
PASS colorComponents[1] is 208
PASS colorComponents[2] is 0
PASS colorComponents[3] is 104

End condition:
PASS colorComponents[1] is 208
PASS colorComponents[2] is 0
PASS colorComponents[3] is 208
PASS successfullyParsed is true

TEST COMPLETE
